Calculating contribution values (capacitive-diffusion)
This tool calculates the k values by calculating the capacitive or diffusive contributions of the data received from the battery tester according to the scanning speed. 
The data taken from Excel is processed in a simple way in the Matlab environment.
The linear regression method was used while calculating the K values. The coefficients of the equation give us the values we are looking for.